Move refactored code in package
Ideally, all or most of the stuff in main.py should be refactored and moved
into the pycraft package over time.
