+++ pycraft/object_registry.py
