+++ pycraft/objects/defaults.png
